Add StudentRestController and related tests
Introduced a new StudentRestController which provides APIs for handling student data. This includes endpoints for fetching a default student, all students, a student by ID, and a student by details. Necessary unit tests are also included in StudentRestControllerTest. A change was also made in ".mvn/extensions.xml" to include a new Maven extension.